update package structure
